# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Kargo",
        "title": "Kargo API",
        "contact": {},
        "version": "v1alpha1"
    },
    "basePath": "/",
    "paths": {
...
        "/v1beta1/projects/{project}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a Project resource.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core",
                    "Cluster-Scoped Resource"
                ],
                "summary": "Retrieve a Project resource",
                "operationId": "GetProject",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project custom resource (github.com/akuity/kargo/api/v1alpha1.Project)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a Project resource and its associated namespace.",
                "tags": [
                    "Core",
                    "Cluster-Scoped Resource"
                ],
                "summary": "Delete a Project",
                "operationId": "DeleteProject",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                }
            }
        },
...
        "/v1beta1/projects/{project}/repo-credentials": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List project-level repository credentials. Returns a SecretList\nresource containing heavily redacted Secrets.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credentials",
                    "Repo Credentials",
                    "Project-Level"
                ],
                "summary": "List project-level repository credentials",
                "operationId": "ListProjectRepoCredentials",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SecretList resource (k8s.io/api/core/v1.SecretList)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create project-level repository credentials. Returns a heavily\nredacted Kubernetes Secret resource.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credentials",
                    "Repo Credentials",
                    "Project-Level"
                ],
                "summary": "Create project-level repository credentials",
                "operationId": "CreateProjectRepoCredentials",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateRepoCredentialsRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Secret resource (k8s.io/api/core/v1.Secret)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/v1beta1/shared/repo-credentials": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List shared repository credentials. Returns a SecretList\nresource containing heavily redacted Secrets.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credentials",
                    "Repo Credentials",
                    "Shared"
                ],
                "summary": "List shared repository credentials",
                "operationId": "ListSharedRepoCredentials",
                "responses": {
                    "200": {
                        "description": "SecretList resource (k8s.io/api/core/v1.SecretList)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create shared repository credentials. Returns a heavily\nredacted Kubernetes Secret resource.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credentials",
                    "Repo Credentials",
                    "Shared"
                ],
                "summary": "Create shared repository credentials",
                "operationId": "CreateSharedRepoCredentials",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateRepoCredentialsRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Secret resource (k8s.io/api/core/v1.Secret)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CreateRepoCredentialsRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "repoUrl": {
                    "type": "string"
                },
                "repoUrlIsRegex": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token authentication. Obtain token via OIDC/PKCE flow with your identity provider.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}